requirements.txt
python-dotenv
openai == 1.37.0 